requirements.txt
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.26.1
nltk==3.6.3
openai==0.27.0
pandas==1.3.4
requests==2.26.0
scikit-learn==1.0.1
sklearn==0.0
tkinter==8.6.10